backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
PyPDF2==3.0.1
groq==0.8.0
python-dotenv==1.0.1